+++ cli/CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Added support for `@<file>` syntax for reading file list from `<file>`
+
+
 0.1.1
 -----
 - Added `--print-lints` argument for printing a list of built-in lints